+++ oasis-web/src/locales/langs/en-us.ts
@@ -240,6 +240,7 @@ const local: App.I18n.Schema = {
     schedule: 'Scheduler',
     schedule_job: 'Job Manage',
     schedule_log: 'Schedule Logs',
+    schedule_dispatch: 'Dispatch Queue',
     'schedule_job_alarm-events': 'Job Alarm Events'
   },
   page: {