+++ oasis-web/src/typings/elegant-router.d.ts
@@ -31,6 +31,7 @@ declare module "@elegant-router/types" {
     "manage_user": "/manage/user";
     "manage_user-detail": "/manage/user-detail/:id";
     "schedule": "/schedule";
+    "schedule_dispatch": "/schedule/dispatch";
     "schedule_job": "/schedule/job";
     "schedule_job_alarm-events": "/schedule/job/alarm-events/:jobId";
     "schedule_log": "/schedule/log";
@@ -102,6 +103,7 @@ declare module "@elegant-router/types" {
     | "manage_role"
     | "manage_user-detail"
     | "manage_user"
+    | "schedule_dispatch"
     | "schedule_job_alarm-events"
     | "schedule_job"
     | "schedule_log"